+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## Version 2.0.0.dev21 (in development)
 
+* Fix dataset name can contain only letters, numbers underscore and dot [#343](https://github.com/CCI-Tools/cate/issues/343)
 * Fix non-spatial data variables sometimes being altered when using subset_spatial [#786](https://github.com/CCI-Tools/cate/issues/786)
 * Fix subsetting datasets with lon>180 when using open_dataset [#787](https://github.com/CCI-Tools/cate/issues/787)
 * Fix for Aerosol data download error [#664](https://github.com/CCI-Tools/cate/issues/664)